api_requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiofiles==23.2.1

# Azure Blob Storage
azure-storage-blob==12.19.0

# Dolphin module dependencies (reusing existing requirements)
torch==2.1.0
torchvision==0.16.0
transformers==4.47.0
accelerate==1.6.0
pillow==9.3.0
opencv-python==4.11.0.86
numpy==1.24.4
pymupdf==1.26
omegaconf==2.3.0
timm==0.5.4

# Additional utilities
python-multipart==0.0.6
python-dotenv==1.0.0
